DI changes and sudoku id generator

+++ Assets/Content/Features/MainScreenModule/Scripts/ISudokuGridUserIntentEmitter.cs
@@ -2,7 +2,7 @@
 {
     public interface ISudokuGridUserIntentEmitter
     {
-        void RequestEdit(int gridId);
+        void RequestEdit(long gridId);
         void RequestCreation();
     }
 }
